+++ packages/adapters/ethers5/src/client.ts
@@ -302,7 +302,7 @@ export class EVMEthers5Client {
         this.appKit?.setClientId(null)
 
         if (this.options?.siweConfig?.options?.signOutOnDisconnect) {
-          const { SIWEController } = await import('@rerock/siwe')
+          const { SIWEController } = await import('@rerock/appkit-siwe')
           await SIWEController.signOut()
         }