+++ packages/scaffold-ui/src/modal/w3m-modal/index.ts
@@ -107,7 +107,7 @@ export class W3mModal extends LitElement {
     const isApproveSignScreen = RouterController.state.view === 'ApproveTransaction'
 
     if (this.isSiweEnabled) {
-      const { SIWEController } = await import('@rerock/siwe')
+      const { SIWEController } = await import('@rerock/appkit-siwe')
       const isUnauthenticated = SIWEController.state.status !== 'success'
       if (isUnauthenticated && (isSiweSignScreen || isApproveSignScreen)) {
         ModalController.shake()
@@ -198,7 +198,7 @@ export class W3mModal extends LitElement {
     this.caipAddress = caipAddress
 
     if (this.isSiweEnabled) {
-      const { SIWEController } = await import('@rerock/siwe')
+      const { SIWEController } = await import('@rerock/appkit-siwe')
       const session = await SIWEController.getSession()
 
       // If the address has changed and signOnAccountChange is enabled, sign out
@@ -215,7 +215,7 @@ export class W3mModal extends LitElement {
     const prevIsConnected = this.connected
 
     if (nextIsConnected && this.isSiweEnabled) {
-      const { SIWEController } = await import('@rerock/siwe')
+      const { SIWEController } = await import('@rerock/appkit-siwe')
       const session = await SIWEController.getSession()
 
       if (!session) {
@@ -251,7 +251,7 @@ export class W3mModal extends LitElement {
 
     if (prevCaipNetworkId && nextNetworkId && prevCaipNetworkId !== nextNetworkId) {
       if (this.isSiweEnabled) {
-        const { SIWEController } = await import('@rerock/siwe')
+        const { SIWEController } = await import('@rerock/appkit-siwe')
         const session = await SIWEController.getSession()
 
         if (session) {