+++ apps/laboratory/CHANGELOG.md
@@ -16,7 +16,7 @@
   - @rerock/appkit@0.0.3
   - @rerock/appkit-ethers@0.0.3
   - @rerock/appkit-ethers5@0.0.3
-  - @rerock/siwe@0.0.3
+  - @rerock/appkit-siwe@0.0.3
   - @rerock/solana@0.0.3
   - @rerock/wagmi@0.0.3
   - @rerock/wallet@0.0.3