+++ packages/ethers/CHANGELOG.md
@@ -14,5 +14,5 @@
   - @rerock/appkit-common@0.0.3
   - @rerock/appkit-polyfills@0.0.3
   - @rerock/appkit-utils@0.0.3
-  - @rerock/siwe@0.0.3
+  - @rerock/appkit-siwe@0.0.3
   - @rerock/wallet@0.0.3