+++ packages/scaffold-ui/CHANGELOG.md
@@ -12,6 +12,6 @@
   - @rerock/appkit-common@0.0.3
   - @rerock/appkit-core@0.0.3
   - @rerock/appkit-utils@0.0.3
-  - @rerock/siwe@0.0.3
+  - @rerock/appkit-siwe@0.0.3
   - @rerock/ui@0.0.3
   - @rerock/wallet@0.0.3